feat: created dynamic sql query for finding a product

+++ services/barilo-api/internal/agolia/product.go
@@ -3,6 +3,6 @@ package agolia
 type ProductIndex struct {
 	ID    string  `json:"id"`
 	Name  string  `json:"name"`
-	Price float64 `json:"price"`
+	Price uint64  `json:"price"`
 	Brand *string `json:"brand"`
 }
